+++ plugins/golang-filter/mcp-server/internal/redis.go
@@ -1,4 +1,4 @@
-package common
+package internal
 
 import (
 	"context"
