+++ plugins/golang-filter/mcp-server/internal/registry.go
