+++ plugins/golang-filter/mcp-server/internal/server.go
