+++ plugins/golang-filter/mcp-server/internal/sse.go
@@ -1,4 +1,4 @@
-package common
+package internal
 
 import (
 	"encoding/json"
